Crate pallet_salary

Make periodic payment to members of a ranked collective according to rank.

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for pallet_salary

Structs

ClaimantStatus

The status of a single payee/claimant.

StatusType

The status of the pallet instance.

Enums

ClaimState

The state of a specific payment claim.

Type Aliases

Cycle

Payroll cycle.