pallet_salary::pallet

Module dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

bump

Move to next payout cycle, assuming that the present block is now within that cycle.

check_payment

Update a payment’s status; if it failed, alter the state so the payment can be retried.

induct

Induct oneself into the payout system.

init

Start the first payout cycle.

payout

Request a payout.

payout_other

Request a payout to a secondary account.

register

Register for a payout.