Trait pallet_salary::pallet::Config

pub trait Config<I: 'static = ()>: Config {
    type WeightInfo: WeightInfo;
    type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>;
    type Paymaster: Pay<AccountId = <Self as Config>::AccountId>;
    type Members: RankedMembers<AccountId = <Self as Config>::AccountId>;
    type Salary: GetSalary<<Self::Members as RankedMembers>::Rank, Self::AccountId, <Self::Paymaster as Pay>::Balance>;
    type RegistrationPeriod: Get<Self::BlockNumber>;
    type PayoutPeriod: Get<Self::BlockNumber>;
    type Budget: Get<BalanceOf<Self, I>>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>

The runtime event type.

type Paymaster: Pay<AccountId = <Self as Config>::AccountId>

Means by which we can make payments to accounts. This also defines the currency and the balance which we use to denote that currency.

type Members: RankedMembers<AccountId = <Self as Config>::AccountId>

The current membership of payees.

type Salary: GetSalary<<Self::Members as RankedMembers>::Rank, Self::AccountId, <Self::Paymaster as Pay>::Balance>

The maximum payout to be made for a single period to an active member of the given rank.

The benchmarks require that this be non-zero for some rank at most 255.

type RegistrationPeriod: Get<Self::BlockNumber>

The number of blocks within a cycle which accounts have to register their intent to claim.

The number of blocks between sequential payout cycles is the sum of this and PayoutPeriod.

type PayoutPeriod: Get<Self::BlockNumber>

The number of blocks within a cycle which accounts have to claim the payout.

The number of blocks between sequential payout cycles is the sum of this and RegistrationPeriod.

type Budget: Get<BalanceOf<Self, I>>

The total budget per cycle.

This may change over the course of a cycle without any problem.